Pass edd_get_payment_status a Post Object
The edd_get_payment_status function requires a post object - not a post ID. This returns nothing when passed a post ID and thus, doesn't work to show PDF invoices for purchases which are still pending. By getting the post object first, then passing that object to edd_get_payment_status, it returns the proper values. Interestingly, the edd_get_payment_status function doesn't need the whole post object as it only uses the ID. It could be made to do a check for if it got a single ID instead of an object - but that's a separate issue.