Add email template tag override for {billing_address} to use full country/state names
Normally, EDDs {billing_address} template tag uses the country/state shortcodes in the purchase receipt.

For example, if the state was "Alabama" and the country was "United States", it would output "AL, US" instead of the full names. This snippet will remove the normal template tag and replace it with the same tag but a different callback function that outputs the full names.