Gots to support those price ids yo
I'm not sure if this is going to be supported in the future but I think it should.

Also, I'd like to see new params added to the filter, i.e. beta should most definitely be added and it would be nice to have all_files to prevent a duplicate lookup.